requirements.txt
# Core
streamlit>=1.32.0
pandas>=2.2.0
python-dotenv>=1.0.0

# Google Cloud
google-cloud-bigquery>=3.13.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Gemini / Generative AI
google-generativeai>=0.7.0

# Data validation
pydantic>=2.6.0

# Utils
requests>=2.31.0